【back】payment/provider を sub-package 化（interface + data 分離）
